requirements.txt
# Core
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.4.2
scipy==1.11.4
joblib==1.3.2
pyarrow==16.1.0
fastparquet==2024.5.0
python-dotenv==1.0.1
pyyaml==6.0.1
tqdm==4.66.4

# Imbalance & models
imbalanced-learn==0.12.2
xgboost==2.0.3
lightgbm==4.3.0

# Explainability
shap==0.45.1
lime==0.2.0.1

# Fairness
fairlearn==0.10.0

# Apps/plots
streamlit==1.35.0
matplotlib==3.8.4
seaborn==0.13.2
plotly==5.22.0

# Experiment tracking (local)
mlflow==2.12.1

# Kaggle CLI
kaggle==1.6.14

# Small helpers
scikit-plot==0.3.7
